+++ pkg/models/status.go
@@ -4,6 +4,6 @@ package models
 type CmpSetupCheck struct {
 	Success      bool   `json:"success"`
 	BuildVersion string `json:"buildVersion"`
-	ApplianceUrl string `json:"applianceUrl"`
+	ApplianceURL string `json:"applianceUrl"`
 	SetupNeeded  bool   `json:"setupNeeded"`
 }
